+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 * Lives left are now shown as a title instead of the chat
+* Fixed lives remaining lives showing up in creative/spectator and hardcore
 
 ## 1.0.1
 * Fixed "Half hearts' penalty on death" could not be zero or negative